+++ assignment1/ATM_Bank_b.c
@@ -5,9 +5,9 @@ int main()
 {
     // Create variables
     int choice;
-    float balance;
-    float deposit;
-    float withdraw;
+    double balance;
+    double deposit;
+    double withdraw;
     float interest_rate;
     int years;
 
@@ -30,14 +30,14 @@ int main()
         // Deposit
         case 1:
             printf("Enter the amount you would like to deposit: $");
-            scanf("%f", &deposit);
+            scanf("%lf", &deposit);
             balance += deposit;
             break;
 
         // Withdraw
         case 2:
             printf("Enter the amount you would like to withdraw: $");
-            scanf("%f", &withdraw);
+            scanf("%lf", &withdraw);
             
             // Checks is user is trying to withdraw more than their balance
             if (withdraw > balance) {
